fixed illustrated works bug, creator role, and additional organizations issues
